# agent2agent_protocol/api_output/openapi.yaml
openapi: 3.0.3
info:
  title: todo_list_api
  version: 1.0.0
  description: API for managing todo items.
servers:
  - url: /api/v1
tags:
  - name: Todos
    description: Todo management endpoints
paths:
  /todos:
    post:
      summary: Create a new todo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List all todos
      tags:
        - Todos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
  /todos/{id}:
    get:
      summary: Get a todo by ID
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing todo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedTodo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a todo by ID
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread, and vegetables"
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          example: "2023-10-25T14:32:00Z"
    UpdatedTodo:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread, and vegetables"
        completed:
          type: boolean
          example: true
        updated_at:
          type: string
          example: "2023-10-25T15:00:00Z"
    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread, and vegetables"
    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread, and vegetables"
        completed:
          type: boolean
          example: true
    TodoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        count:
          type: integer
          example: 1
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request payload"